requirements.txt
numpy
matplotlib
Pillow
scipy
svgwrite